requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
Django==3.2.9
django-environ==0.8.1
et-xmlfile==1.1.0
openpyxl==3.0.9
psycopg2-binary==2.9.3
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
